requirements.txt
streamlit
google
google-generativeai
python-dotenv
pdf2image
pillow
fpdf
psycopg2
pdfplumber